feat(events): refactor dashboard logic and add create, edit, and detail routes
