+++ backend/src/index.ts
@@ -46,7 +46,6 @@ app.use(
 app.use('/users', userRouter);
 app.use('/events', eventRouter);
 app.use('/auth', authRouter);
-app.use('/dashboard', eventRouter);
 
 // Health check
 app.get('/', (req: Request, res: Response) => {
