+++ frontend/src/app/(private)/events/create/page.tsx
@@ -1,0 +1,10 @@
+export default function CreateEventPage() {
+  return (
+    <div className="w-full min-h-screen bg-gray-50 p-8 flex flex-col items-center justify-center text-center">
+      <h1 className="text-2xl font-bold text-gray-800 uppercase tracking-wide">
+        🛠️ Create New Event
+      </h1>
+      <p className="text-sm text-gray-500 mt-2 font-mono">Route: /events/create</p>
+    </div>
+  );
+}
