+++ frontend/src/app/(public)/sign-in/page.tsx
@@ -13,8 +13,7 @@ export default function LoginPage() {
     const loggedIn = await signIn(data);
 
     if (loggedIn) {
-      //router.push('/events');
-      router.push('/dashboard');
+      router.push('/events');
     }
   };
