+++ frontend/src/app/(public)/sign-up/page.tsx
@@ -18,8 +18,7 @@ const SignUp = () => {
       await registerRequest(data);
       // After successful registration, log the user in and redirect to /events
       await loginRequest(data.email, data.password);
-      //router.push('/events');
-      router.push('/dashboard');
+      router.push('/events');
     } catch (err: unknown) {
       if (err instanceof Error) {
         setServerError(err.message);
